LayoutTests/inspector/canvas/create-context-bitmaprenderer.html

<!DOCTYPE html>
<html>
<head>
<link rel="stylesheet" href="resources/create-context-utilities.css">
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="resources/create-context-utilities.js"></script>
<script>
function test() {
    let suite = initializeTestSuite("Canvas.CreateContextBitmapRenderer");

    addSimpleTestCase({
        name: "Attached",
        expression: `createAttachedCanvas("bitmaprenderer")`,
        contextType: WI.Canvas.ContextType.BitmapRenderer,
    });

    addSimpleTestCase({
        name: "Detached",
        expression: `createDetachedCanvas("bitmaprenderer")`,
        contextType: WI.Canvas.ContextType.BitmapRenderer,
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
    <p>Test that CanvasManager tracks creation and destruction of BitmapRenderer canvases.</p>
</body>
</html>
