LayoutTests/inspector/canvas/create-context-bitmaprenderer-expected.txt

Test that CanvasManager tracks creation and destruction of BitmapRenderer canvases.


== Running test suite: Canvas.CreateContextBitmapRenderer
-- Running test case: Canvas.CreateContextBitmapRenderer.NoCanvases
PASS: CanvasManager should have no canvases.

-- Running test case: Canvas.CreateContextBitmapRenderer.Attached
PASS: Canvas context should be Bitmap Renderer.
  0: getContext - [native code]
  1: createAttachedCanvas - inspector/canvas/resources/create-context-utilities.js:4:36
  2: Global Code - [program code]
  3: evaluateWithScopeExtension - [native code]
  4: (anonymous function) - [native code]
  5: _wrapCall - [native code]

PASS: Removed canvas has expected ID.

-- Running test case: Canvas.CreateContextBitmapRenderer.Detached
PASS: Canvas context should be Bitmap Renderer.
  0: getContext - [native code]
  1: createDetachedCanvas - inspector/canvas/resources/create-context-utilities.js:8:62
  2: Global Code - [program code]
  3: evaluateWithScopeExtension - [native code]
  4: (anonymous function) - [native code]
  5: _wrapCall - [native code]

PASS: Removed canvas has expected ID.
